libcaer: change in-memory format. Use all the bits in EAR events. Increase possible channels in FRAME to 128. Widen info to 32bit for better memory layout in IMU6/9. Increase X/Y address for POLARITY to 15bits, using all bits.
git-svn-id: https://svn.code.sf.net/p/jaer/code/libcaer/trunk@7229 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ include/events/imu6.h
@@ -11,7 +11,7 @@
 #include "common.h"
 
 struct caer_imu6_event {
-	uint16_t info; // First because of valid mark.
+	uint32_t info; // First because of valid mark.
 	int32_t timestamp;
 	float accel_x;
 	float accel_y;
